+++ __tests__/unit/block_layout/mixed_layout.test.ts
@@ -1,4 +1,4 @@
-import { Settings } from "../../../src/libs/engine/layout_engine/settings";
+import { Settings } from "../../../src/libs/engine/settings";
 import { LayoutDirector } from "../../../src/libs/engine/layout_engine/builder/layout_director";
 import { Alignment } from "../../../src/libs/common/alignment.enum";
 import { LayoutTree } from "../../../src/libs/engine/layout_engine/builder/groups/layout_tree";
