+++ __tests__/unit/complete_layout/nested_layout.test.ts
@@ -2,7 +2,7 @@ import { LayoutTypes } from "../../../src/libs/common/layout_types.enum";
 import { SmartViewEngine } from "../../../src";
 import { View } from "../../../src/libs/model/view";
 import { HydratedViewNode, ViewNode } from "../../../src/libs/model/view_node";
-import { Settings } from "../../../src/libs/engine/layout_engine/settings";
+import { Settings } from "../../../src/libs/engine/settings";
 import { LayoutDirector } from "../../../src/libs/engine/layout_engine/builder/layout_director";
 import { Alignment } from "../../../src/libs/common/alignment.enum";
