+++ src/index.ts
@@ -1,3 +1,4 @@
 import { SmartViewEngine } from "@libs/engine/smart_view_engine";
+import { Settings } from "@libs/engine/settings";
 
-export { SmartViewEngine };
+export { SmartViewEngine, Settings };
