+++ src/libs/engine/layout_engine/builder/groups/layout_tree.ts
@@ -1,4 +1,4 @@
-import { Settings } from "@libs/engine/layout_engine/settings";
+import { Settings } from "@libs/engine/settings";
 import { LayoutGroup } from "@libs/engine/layout_engine/builder/groups/layout_group";
 import { Alignment } from "@libs/common/alignment.enum";
 import { BaseElement } from "@libs/model/base_element";
