+++ src/libs/engine/layout_engine/builder/groups/visible_layout_row.ts
@@ -1,5 +1,5 @@
 import { Alignment } from "@libs/common/alignment.enum";
-import { Settings } from "@libs/engine/layout_engine/settings";
+import { Settings } from "@libs/engine/settings";
 import { LayoutRow } from "@libs/engine/layout_engine/builder/groups/layout_row";
 import { Direction } from "@libs/common/distribution.enum";
 import { ContentBox } from "@libs/engine/layout_engine/builder/groups/content_box/content_box";
