+++ src/libs/engine/layout_engine/linear/hierarchical/hierarchical_layout_engine.ts
@@ -2,7 +2,7 @@ import { LayoutEngine } from "../../layout_engine";
 import { LayoutDirector } from "@libs/engine/layout_engine/builder/layout_director";
 import { Alignment } from "@libs/common/alignment.enum";
 import { HydratedViewNode } from "@libs/model/view_node";
-import { Settings } from "@libs/engine/layout_engine/settings";
+import { Settings } from "@libs/engine/settings";
 import { SemanticEngine } from "@libs/engine/semantic_engine/semantic_engine";
 import { HydratedView } from "@libs/model/hydrated_view";
