+++ src/libs/engine/layout_engine/linear/hierarchical/specialized/nested_layout_engine.ts
@@ -1,5 +1,5 @@
 import { HierarchicalLayoutEngine } from "../hierarchical_layout_engine";
-import { Settings } from "@libs/engine/layout_engine/settings";
+import { Settings } from "@libs/engine/settings";
 import { SemanticEngine } from "@libs/engine/semantic_engine/semantic_engine";
 import { HydratedViewNode } from "@libs/model/view_node";
 import { LayoutDirector } from "@libs/engine/layout_engine/builder/layout_director";
