+++ src/libs/engine/smart_view_engine.ts
@@ -1,5 +1,5 @@
 import { LayoutTypes } from "../common/layout_types.enum";
-import { Settings } from "@libs/engine/layout_engine/settings";
+import { Settings } from "@libs/engine/settings";
 import { NestedLayoutEngine } from "./layout_engine/linear/hierarchical/specialized/nested_layout_engine";
 import { HierarchyLayoutEngine } from "./layout_engine/linear/hierarchical/specialized/hierarchy_layout_engine";
 import { SemanticEngine } from "./semantic_engine/semantic_engine";
